+++ src/adapter/zoh/adapter/utils.ts
@@ -11,7 +11,7 @@ export function bigUInt64ToHexBE(value: bigint): string {
  * @returns 8-bytelength buffer in little-endian
  */
 export function bigUInt64ToBufferLE(value: bigint): Buffer {
-    const b = Buffer.alloc(8);
+    const b = Buffer.allocUnsafe(8);
     b.writeBigUInt64LE(value, 0);
     return b;
 }
@@ -21,7 +21,7 @@ export function bigUInt64ToBufferLE(value: bigint): Buffer {
  * @returns 8-bytelength buffer in big-endian
  */
 export function bigUInt64ToBufferBE(value: bigint): Buffer {
-    const b = Buffer.alloc(8);
+    const b = Buffer.allocUnsafe(8);
     b.writeBigUInt64BE(value, 0);
     return b;
 }